Bootloader_v1
Made the most basic bootloader that jumps to the main application. Added required padding to the bootloader executable to make sure that it always occupies 64KB. Added the bootloader section to the linker script of the main application, so that the bootloader remains at the top of the memory, followed by the vector table of the main application and then the main application itself. Used the Vector Table Offset register(VTOR, from the scb header file) to ensure that the interrupts refer to the correct vector table for respective handlers.