Bootloader_v2
Using the "arm-none-eabi-objdump -D bootloader.elf > asm-dump", one can see the generated assembly file. Having a look at the assembly of the corresponding C files(in this case bootloader.c) can prove to be helpful at times.
In this version of the bootloader, I have slightly modified how we access the reset vector. This is a slightly better way of doing it, but note that a pointer can't always be treated as an array.