Bootloader_v3
The most appealing and probably the best way to make it. Made use of the 'vector.h' header file. Used the concept of accessing a member of the struct using a pointer. Note that all the three asm_dump files have the same assembly instructions for the main function, hence confirming the fact that all the three versions of the bootloader are effectively the same.